OJ-1921: Moved `nino-check` tests to `nino-check` folder
