+++ integration-tests/tests/aws/nino-check/nino-check-unhappy.test.ts
@@ -1,6 +1,6 @@
-import { stackOutputs } from "./resources/cloudformation-helper";
-import { clearItems, populateTable } from "./resources/dynamodb-helper";
-import { executeStepFunction } from "./resources/stepfunction-helper";
+import { stackOutputs } from "../resources/cloudformation-helper";
+import { clearItems, populateTable } from "../resources/dynamodb-helper";
+import { executeStepFunction } from "../resources/stepfunction-helper";
 
 describe("Nino Check", () => {
   const sessionId = "12345";
